refactor(threads): use single thread update chan on wallet
